rename activitytag package
