+++ src/main/java/com/asialocalguide/gateway/activitytag/controller/ActivityTagController.java
@@ -1,12 +1,13 @@
-package com.asialocalguide.gateway.activity.controller;
-
-import com.asialocalguide.gateway.activity.service.ActivityTagService;
-import com.asialocalguide.gateway.core.dto.activitytag.ActivityTagDTO;
+package com.asialocalguide.gateway.activitytag.controller;
 
 import java.util.List;
+
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.asialocalguide.gateway.activitytag.dto.ActivityTagDTO;
+import com.asialocalguide.gateway.activitytag.service.ActivityTagService;
 
 @RestController
 @RequestMapping("/v1/activity-tags")
