+++ src/main/java/com/asialocalguide/gateway/activitytag/domain/ActivityTagProviderMappingId.java
@@ -1,9 +1,10 @@
-package com.asialocalguide.gateway.activity.domain;
+package com.asialocalguide.gateway.activitytag.domain;
+
+import java.io.Serializable;
+import java.util.Objects;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import java.io.Serializable;
-import java.util.Objects;
 import lombok.Getter;
 
 @Embeddable
