+++ src/main/java/com/asialocalguide/gateway/activitytag/dto/ActivityTagDTO.java
@@ -1,4 +1,4 @@
-package com.asialocalguide.gateway.core.dto.activitytag;
+package com.asialocalguide.gateway.activitytag.dto;
 
 public record ActivityTagDTO(Long id, String name, String promptText) {
 }
