+++ src/main/java/com/asialocalguide/gateway/activitytag/repository/ActivityTagRepository.java
@@ -1,0 +1,9 @@
+package com.asialocalguide.gateway.activitytag.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.asialocalguide.gateway.activitytag.domain.ActivityTag;
+import com.asialocalguide.gateway.core.repository.custom.CustomActivityTagRepository;
+
+public interface ActivityTagRepository extends JpaRepository<ActivityTag, Long>, CustomActivityTagRepository {
+}
