--- src/main/java/com/asialocalguide/gateway/core/repository/ActivityTagRepository.java
@@ -1,8 +1,0 @@
-package com.asialocalguide.gateway.core.repository;
-
-import com.asialocalguide.gateway.activity.domain.ActivityTag;
-import com.asialocalguide.gateway.core.repository.custom.CustomActivityTagRepository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ActivityTagRepository extends JpaRepository<ActivityTag, Long>, CustomActivityTagRepository {
-}
